fully restructure entire codebase, turn command modules into extensions fixes and improvements
