+++ Procfile
@@ -1,1 +1,1 @@
-worker: python3 main.py
+worker: python3 -m pgbot
