--- bootstrap.json
@@ -1,13 +1,0 @@
-{
-    "extensions": [
-        {
-            "name": "pgbot.commands.help"
-        },
-        {
-            "name": "pgbot.commands.admin"
-        },
-        {
-            "name": "pgbot.commands.user"
-        }
-    ]
-}
