+++ pgbot/__main__.py
@@ -1,0 +1,12 @@
+"""
+This file is a part of the source code for the PygameCommunityBot.
+This project has been licensed under the MIT license.
+Copyright (c) 2020-present pygame-community
+
+This file is the main file of the PygameCommunityBot source. Running this
+starts the bot
+"""
+
+import pgbot
+
+pgbot.run()
