+++ pgbot/exceptions.py
@@ -3,7 +3,7 @@ from discord.ext import commands
 """
 This file is a part of the source code for the PygameCommunityBot.
 This project has been licensed under the MIT license.
-Copyright (c) 2020-present PygameCommunityDiscord
+Copyright (c) 2020-present pygame-community
 
 This file defines some exceptions used across the whole codebase
 """
