+++ pgbot/exts/core_commands/admin/sudo.py
@@ -1,7 +1,7 @@
 """
 This file is a part of the source code for the PygameCommunityBot.
 This project has been licensed under the MIT license.
-Copyright (c) 2020-present PygameCommunityDiscord
+Copyright (c) 2020-present pygame-community
 
 This file defines the command handler class for the sudo commands of the bot
 """
@@ -22,10 +22,9 @@ import snakecore
 
 from pgbot import common
 import pgbot
-from pgbot.commands.base import BaseCommandCog
-from pgbot.commands.utils.checks import admin_only_and_custom_parsing
-from pgbot.commands.utils.cogs import CommandUtilsCog
-from pgbot.commands.utils.converters import String
+from ..base import BaseCommandCog
+from ..utils.checks import admin_only_and_custom_parsing
+from ..utils.converters import String
 from pgbot.exceptions import BotException
 
 process = psutil.Process(os.getpid())
@@ -648,7 +647,7 @@ class SudoCommandCog(BaseCommandCog):
                 filesize_limit = (
                     ctx.guild.filesize_limit
                     if ctx.guild is not None
-                    else common.BASIC_MAX_FILE_SIZE
+                    else common.DEFAULT_FILESIZE_LIMIT
                 )
                 with io.StringIO("This file was too large to be duplicated.") as fobj:
                     attached_files = [
@@ -1118,7 +1117,7 @@ class SudoCommandCog(BaseCommandCog):
                 filesize_limit = (
                     ctx.guild.filesize_limit
                     if ctx.guild is not None
-                    else common.BASIC_MAX_FILE_SIZE
+                    else common.DEFAULT_FILESIZE_LIMIT
                 )
                 with io.StringIO("This file was too large to be cloned.") as fobj:
                     attached_files = [
