+++ pgbot/exts/core_commands/utils/__init__.py
@@ -1,7 +1,7 @@
 """
 This file is a part of the source code for the PygameCommunityBot.
 This project has been licensed under the MIT license.
-Copyright (c) 2020-present PygameCommunityDiscord
+Copyright (c) 2020-present pygame-community
 
 This module defines some helpful utilities for bot commands.
 """
@@ -12,5 +12,4 @@ from discord.ext import commands
 
 from pgbot import common
 
-from . import checks, clock, cogs, docs, help, sandbox
-from .utils import *
+from . import checks, clock, docs, help, sandbox
