+++ requirements.txt
@@ -6,5 +6,5 @@ psutil>=5.7.3
 pygame>=2.0.0
 pygame_gui>=0.5.7
 python-dotenv>=0.17
-snakecore @ git+https://github.com/PygameCommunityDiscord/snakecore.git@main#egg=snakecore-0.1
+snakecore @ git+https://github.com/pygame-community/snakecore.git@main#egg=snakecore-0.1
 unidecode>=1.2.0
