+++ setup.cfg
@@ -1,5 +1,5 @@
 [flake8]
-max-line-length = 80
+max-line-length = 88
 max-complexity = 12
 
 ignore = E203,E501,W503
